+++ lesson_04/interfaces.go
@@ -1,0 +1,6 @@
+package main
+
+type Shape interface {
+	Area() float64
+	Perimeter() float64
+}
